Problems with update and delete on all.  ALso problems with post garment and adding fabric and pattern

+++ src/main/java/com/promineotech/sewingProject/service/NotebookService.java
@@ -1,10 +1,14 @@
 package com.promineotech.sewingProject.service;
 
+import java.util.Optional;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.promineotech.sewingProject.entity.Fabric;
+import com.promineotech.sewingProject.entity.Garment;
 import com.promineotech.sewingProject.entity.Notebook;
 import com.promineotech.sewingProject.entity.User;
 import com.promineotech.sewingProject.repository.NotebookRepository;
@@ -32,18 +36,36 @@ public class NotebookService {
 	}
 	
 	//get all notebooks
-	public Iterable<Notebook> getAllNotebooks(){
-		return repo.findAll();
+	public Iterable<Notebook> getNotebooksByUser(Long userId){
+		return (repo.findByUserId(userId));
 	}
 	
 	//get particular notebook
-	public Notebook getNotebook(Long id) {
-		return repo.findById(id).get();
+	public Notebook getNotebook(Long userId, Long id) {
+		//return repo.findById(id).get();
+		Optional<Notebook> responseNotebook = repo.findById(id);
+		if (responseNotebook.isPresent()) {
+			Notebook notebook = responseNotebook.get();
+			if (notebook.getId() == id) {
+				User user = notebook.getUser();
+				if (user.getId() == userId) {
+					return (notebook);
+				}
+			}
+		}
+		return (null);
 	}
 	
 	//update notebook
 	
-	public Notebook updateNotebook(Notebook notebook, Long id) throws Exception {
+	//the problem here is that it will update notebook regardless of if the userid in the path is correct or not.
+	//I'm not sure how to validate the userId to make the code run.
+	public Notebook updateNotebook(Notebook notebook, Long id, Long userId) throws Exception {
+		User user = userRepo.findById(userId).get();
+		//if (user == null) {
+		//	throw new Exception("User not found.");
+		//}
+		if (user.getId() == userId) {
 		try {
 			Notebook oldNotebook = repo.findById(id).get();
 			oldNotebook.setName(notebook.getName());
@@ -53,10 +75,16 @@ public class NotebookService {
 			logger.error("Exception occurrec while trying to update notebook: " + id, e);
 			throw new Exception("Unable to update notebook.");
 		}
+		}
+		return null;
 	}
 
-	//delete notebook	
-	public void deleteNotebook(Long id) {
+	//delete notebook	this doesn't work, will delete regardless of userId
+	public void deleteNotebook(Long id, Long userId) throws Exception {
+		User user = userRepo.findById(userId).get();
+		if (user == null) {
+			throw new Exception("User not found.");
+		}
 		repo.deleteById(id);
 	}
 }
